pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "flask>=3.1.2",
    "lxml>=5.4.0",
    "pandas>=2.3.2",
    "playwright>=1.54.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
]
